+++ src/common/List.hpp
@@ -7,7 +7,7 @@
 
 namespace dab::detail::common {
 
-template <typename T, size_t Cap>
+template <typename T, uint32_t Cap>
 class List : public Iterable<List<T, Cap>> {
   public:
   constexpr List() = default;
@@ -25,7 +25,7 @@ class List : public Iterable<List<T, Cap>> {
   constexpr void
   Append(const T item);
 
-  constexpr size_t
+  constexpr uint32_t
   Size() const;
   constexpr T*
   Begin();
@@ -38,54 +38,54 @@ class List : public Iterable<List<T, Cap>> {
 
   private:
   Array<T, Cap> Data;
-  size_t Length = 0;
+  uint32_t Length = 0;
 };
 
-template <typename T, size_t Cap>
+template <typename T, uint32_t Cap>
 constexpr void
 List<T, Cap>::ClearAndSet(const T item) {
   Data[0] = item;
   Length = 1;
 }
 
-template <typename T, size_t Cap>
+template <typename T, uint32_t Cap>
 constexpr void
 List<T, Cap>::Clear() {
   Length = 0;
 }
 
-template <typename T, size_t Cap>
+template <typename T, uint32_t Cap>
 constexpr void
 List<T, Cap>::Append(const T item) {
   assert(Length < Cap);
   Data[Length++] = item;
 }
 
-template <typename T, size_t Cap>
-constexpr size_t
+template <typename T, uint32_t Cap>
+constexpr uint32_t
 List<T, Cap>::Size() const {
   return Length;
 }
 
-template <typename T, size_t Cap>
+template <typename T, uint32_t Cap>
 constexpr T*
 List<T, Cap>::Begin() {
   return Data.Begin();
 }
 
-template <typename T, size_t Cap>
+template <typename T, uint32_t Cap>
 constexpr const T*
 List<T, Cap>::Begin() const {
   return Data.Begin();
 }
 
-template <typename T, size_t Cap>
+template <typename T, uint32_t Cap>
 constexpr T*
 List<T, Cap>::End() {
   return Data.Begin() + Length;
 }
 
-template <typename T, size_t Cap>
+template <typename T, uint32_t Cap>
 constexpr const T*
 List<T, Cap>::End() const {
   return Data.Begin() + Length;
