+++ src/robot/ParallelSearchRobot.hpp
@@ -31,8 +31,8 @@ ParallelSearchRobot<BoardSize, SubRobotSearchTime, SubRobotNumber>::BestCandidat
 
   SearchResult.Reset();
 
-  tbb::parallel_for(tbb::blocked_range<size_t>(0, SubRobots.Size()), [&](const tbb::blocked_range<size_t>& r) {
-    for (size_t i = r.begin(); i != r.end(); ++i) {
+  tbb::parallel_for(tbb::blocked_range<uint32_t>(0, SubRobots.Size()), [&](const tbb::blocked_range<uint32_t>& r) {
+    for (uint32_t i = r.begin(); i != r.end(); ++i) {
       SubRobots[i].BestCandidateEdges(board);
     }
   });
